Factoring out OpenEJB base code to base class, and making service tests simpler
