+++ service/src/test/java/org/jboss/aerogear/unifiedpush/service/PushApplicationServiceTest.java
@@ -16,89 +16,18 @@
  */
 package org.jboss.aerogear.unifiedpush.service;
 
-import org.apache.openejb.jee.Beans;
-import org.apache.openejb.junit.ApplicationComposer;
-import org.apache.openejb.mockito.MockitoInjector;
-import org.apache.openejb.testing.MockInjector;
-import org.apache.openejb.testing.Module;
 import org.jboss.aerogear.unifiedpush.api.PushApplication;
-import org.jboss.aerogear.unifiedpush.dao.InstallationDao;
-import org.jboss.aerogear.unifiedpush.jpa.dao.impl.JPAInstallationDao;
-import org.jboss.aerogear.unifiedpush.jpa.dao.impl.JPAPushApplicationDao;
-import org.jboss.aerogear.unifiedpush.jpa.dao.impl.JPAPushMessageInformationDao;
-import org.jboss.aerogear.unifiedpush.jpa.dao.impl.JPAVariantDao;
-import org.jboss.aerogear.unifiedpush.service.impl.PushApplicationServiceImpl;
-import org.jboss.aerogear.unifiedpush.service.impl.PushSearchByDeveloperServiceImpl;
-import org.jboss.aerogear.unifiedpush.service.impl.PushSearchServiceImpl;
-import org.jboss.aerogear.unifiedpush.service.impl.SearchManager;
-import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.keycloak.KeycloakPrincipal;
-import org.keycloak.KeycloakSecurityContext;
-import org.keycloak.representations.AccessToken;
-import org.mockito.Mock;
 
-import javax.inject.Inject;
-import javax.servlet.http.HttpServletRequest;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.when;
 
-
-@RunWith(ApplicationComposer.class)
 public class PushApplicationServiceTest extends AbstractBaseServiceTest {
 
-    @Inject
-    private PushApplicationServiceImpl pushApplicationService;
-
-    @Inject
-    private PushSearchByDeveloperServiceImpl searchApplicationService;
-
-    @Inject
-    private SearchManager searchManager;
-
-    @Mock
-    private HttpServletRequest httpServletRequest;
-
-    @Mock
-    private KeycloakSecurityContext context;
-
-    @Mock
-    private KeycloakPrincipal keycloakPrincipal;
-
-    @MockInjector
-    public Class<?> mockitoInjector() {
-        return MockitoInjector.class;
-    }
-
-    @Before
-    public void setUp(){
-        AccessToken token = new AccessToken();
-        //The current developer will always be the admin in this testing scenario
-        token.setPreferredUsername("admin");
-        when(context.getToken()).thenReturn(token);
-        when(keycloakPrincipal.getKeycloakSecurityContext()).thenReturn(context);
-        when(httpServletRequest.getUserPrincipal()).thenReturn(keycloakPrincipal);
-        searchManager.setHttpServletRequest(httpServletRequest);
-    }
-
-    @Module
-    public Beans getBeans() {
-        final Beans beans = new Beans();
-        beans.addManagedClass(JPAVariantDao.class);
-        beans.addManagedClass(JPAInstallationDao.class);
-        beans.addManagedClass(JPAPushMessageInformationDao.class);
-        beans.addManagedClass(PushApplicationServiceImpl.class);
-        beans.addManagedClass(JPAPushApplicationDao.class);
-        beans.addManagedClass(PushSearchByDeveloperServiceImpl.class);
-        beans.addManagedClass(PushSearchServiceImpl.class);
-        beans.addManagedClass(SearchManager.class);
-        beans.addManagedClass(InstallationDao.class);
-
-
-        return beans;
+    @Override
+    protected void specificSetup() {
+        // noop
     }
 
     @Test
